Moved utils.utils functions closer to callers.
