+++ environment.yml
@@ -21,6 +21,7 @@ dependencies:
   - main::mkl
   - pip:
     - git+https://github.com/MiraGeoscience/simpeg.git@release/geoapps-0.8.0
+    - git+https://github.com/MiraGeoscience/simpeg.git@v0.9.1.dev1+geoapps.0.6.0
     - git+https://github.com/MiraGeoscience/geoh5py.git@release/0.3.0
     - tqdm
