+++ geoapps/__init__.py
@@ -5,7 +5,7 @@
 #  geoapps is distributed under the terms and conditions of the MIT License
 #  (see LICENSE file at the root of this source code package).
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 from .base.application import BaseApplication
 from .base.plot import PlotSelection2D
