+++ geoapps/interpolation/application.py
@@ -9,18 +9,17 @@ import numpy as np
 from discretize.utils import mesh_utils
 from geoh5py.objects import BlockModel, ObjectBase
 from geoh5py.workspace import Workspace
+from scipy.interpolate import LinearNDInterpolator
+from scipy.spatial import cKDTree
 
+from geoapps.base.selection import ObjectDataSelection, TopographyOptions
+from geoapps.shared_utils.utils import get_locations, weighted_average
 from geoapps.utils import soft_import
 
 (Dropdown, FloatText, HBox, Label, RadioButtons, Text, VBox) = soft_import(
     "ipywidgets",
     objects=["Dropdown", "FloatText", "HBox", "Label", "RadioButtons", "Text", "VBox"],
 )
-from scipy.interpolate import LinearNDInterpolator
-from scipy.spatial import cKDTree
-
-from geoapps.base.selection import ObjectDataSelection, TopographyOptions
-from geoapps.utils.utils import get_locations, weighted_average
 
 
 class DataInterpolation(ObjectDataSelection):
