+++ geoapps/inversion/components/data.py
@@ -21,7 +21,8 @@ from SimPEG import maps
 from SimPEG.electromagnetics.static.utils.static_utils import geometric_factor
 from SimPEG.utils.drivers import create_nested_mesh
 
-from geoapps.utils import calculate_2D_trend, filter_xy
+from geoapps.inversion.utils import calculate_2D_trend
+from geoapps.utils import filter_xy
 
 from .factories import (
     EntityFactory,
