+++ geoapps/inversion/electricals/application.py
@@ -58,7 +58,8 @@ from geoapps import PlotSelection2D
 from geoapps.base.selection import ObjectDataSelection, TopographyOptions
 from geoapps.inversion.electricals import DirectCurrentParams, InducedPolarizationParams
 from geoapps.inversion.electricals.direct_current.constants import app_initializer
-from geoapps.utils.utils import find_value, string_2_list
+from geoapps.utils.general import find_value
+from geoapps.utils.io import string_2_list
 
 
 def inversion_defaults():
@@ -892,8 +893,6 @@ class InversionApp(PlotSelection2D):
             data_channel_options[key].children[3].children[1].options = [
                 ["", None]
             ] + options
-
-            # data_channel_options[key].children[1].value = find_value(options, [key])
 
         self.data_channel_choices.value = inversion_defaults()["component"][
             self.inversion_type.value
