+++ geoapps/inversion/utils.py
@@ -1,0 +1,98 @@
+#  Copyright (c) 2022 Mira Geoscience Ltd.
+#
+#  This file is part of geoapps.
+#
+#  geoapps is distributed under the terms and conditions of the MIT License
+#  (see LICENSE file at the root of this source code package).
+
+import numpy as np
+from scipy.spatial import ConvexHull
+
+
+def calculate_2D_trend(
+    points: np.ndarray, values: np.ndarray, order: int = 0, method: str = "all"
+):
+    """
+    detrend2D(points, values, order=0, method='all')
+
+    Function to remove a trend from 2D scatter points with values
+
+    Parameters:
+    ----------
+
+    points: array or floats, shape(*, 2)
+        Coordinates of input points
+
+    values: array of floats, shape(*,)
+        Values to be de-trended
+
+    order: Order of the polynomial to be used
+
+    method: str
+        Method to be used for the detrending
+            "all": USe all points
+            "perimeter": Only use points on the convex hull
+
+
+    Returns
+    -------
+
+    trend: array of floats, shape(*,)
+        Calculated trend
+
+    coefficients: array of floats, shape(order+1)
+        Coefficients for the polynomial describing the trend
+
+        trend = c[0] + points[:, 0] * c[1] +  points[:, 1] * c[2]
+    """
+    if not isinstance(order, int) or order < 0:
+        raise ValueError(
+            "Polynomial 'order' should be an integer > 0. "
+            f"Value of {order} provided."
+        )
+
+    ind_nan = ~np.isnan(values)
+    loc_xy = points[ind_nan, :]
+    values = values[ind_nan]
+
+    if method == "perimeter":
+        hull = ConvexHull(loc_xy[:, :2])
+        # Extract only those points that make the ConvexHull
+        loc_xy = loc_xy[hull.vertices, :2]
+        values = values[hull.vertices]
+    elif not method == "all":
+        raise ValueError(
+            "'method' must be either 'all', or 'perimeter'. " f"Value {method} provided"
+        )
+
+    # Compute center of mass
+    center_x = np.sum(loc_xy[:, 0] * np.abs(values)) / np.sum(np.abs(values))
+    center_y = np.sum(loc_xy[:, 1] * np.abs(values)) / np.sum(np.abs(values))
+
+    polynomial = []
+    xx, yy = np.triu_indices(order + 1)
+    for x, y in zip(xx, yy):
+        polynomial.append(
+            (loc_xy[:, 0] - center_x) ** float(x)
+            * (loc_xy[:, 1] - center_y) ** float(y - x)
+        )
+    polynomial = np.vstack(polynomial).T
+
+    if polynomial.shape[0] <= polynomial.shape[1]:
+        raise ValueError(
+            "The number of input values must be greater than the number of coefficients in the polynomial. "
+            f"Provided {polynomial.shape[0]} values for a {order}th order polynomial with {polynomial.shape[1]} coefficients."
+        )
+
+    params, _, _, _ = np.linalg.lstsq(polynomial, values, rcond=None)
+    data_trend = np.zeros(points.shape[0])
+    for count, (x, y) in enumerate(zip(xx, yy)):
+        data_trend += (
+            params[count]
+            * (points[:, 0] - center_x) ** float(x)
+            * (points[:, 1] - center_y) ** float(y - x)
+        )
+    print(
+        f"Removed {order}th order polynomial trend with mean: {np.mean(data_trend):.6g}"
+    )
+    return data_trend, params
