+++ geoapps/peak_finder/utils.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import numpy as np
 from scipy.interpolate import interp1d
 
-from geoapps.utils.utils import running_mean
+from geoapps.driver_base.utils import running_mean
 
 
 class LineDataDerivatives:
