+++ geoapps/shared_utils/__init__.py
@@ -4,5 +4,3 @@
 #
 #  geoapps is distributed under the terms and conditions of the MIT License
 #  (see LICENSE file at the root of this source code package).
-
-from .importing import soft_import
