--- geoapps/utils/workspace_objects.py
@@ -1,32 +1,0 @@
-#  Copyright (c) 2022 Mira Geoscience Ltd.
-#
-#  This file is part of geoapps.
-#
-#  geoapps is distributed under the terms and conditions of the MIT License
-#  (see LICENSE file at the root of this source code package).
-
-from __future__ import annotations
-
-from typing import UUID
-from uuid import UUID
-
-
-class WorkspaceObject:
-    def __init__(self, source, target, param):
-        self.source = source
-        self.target = target
-        self.param = param
-
-    def fetch(self, p: str | UUID):
-        """Fetch the object addressed by uuid from the workspace."""
-
-        if isinstance(p, str):
-            try:
-                p = UUID(p)
-            except:
-                p = self.params.__getattribute__(p)
-
-        try:
-            return self.workspace.get_entity(p)[0].values
-        except AttributeError:
-            return self.workspace.get_entity(p)[0]
