+++ tests/run_tests/octree_creation_test.py
@@ -15,7 +15,7 @@ from geoh5py.workspace import Workspace
 from scipy import spatial
 
 from geoapps.octree_creation.application import OctreeMesh
-from geoapps.utils.utils import treemesh_2_octree
+from geoapps.shared_utils.utils import treemesh_2_octree
 
 # pytest.skip("eliminating conflicting test.", allow_module_level=True)
